Extract AuthHttpService
