+++ frontend/src/app.tsx
@@ -3,6 +3,7 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { AppRouter } from './router';
 import { AuthDataServiceContextProvider } from './services/auth-data-service';
+import { AuthHttpServiceContextProvider } from './services/auth.http-service';
 import { HttpServiceContextProvider } from './services/http-service';
 import { ServiceDocsHttpServiceContextProvider } from './services/service-docs.http-service';
 
@@ -11,9 +12,11 @@ export const App: React.FC = () => {
     <BrowserRouter>
       <AuthDataServiceContextProvider>
         <HttpServiceContextProvider>
-          <ServiceDocsHttpServiceContextProvider>
-            <AppRouter />
-          </ServiceDocsHttpServiceContextProvider>
+          <AuthHttpServiceContextProvider>
+            <ServiceDocsHttpServiceContextProvider>
+              <AppRouter />
+            </ServiceDocsHttpServiceContextProvider>
+          </AuthHttpServiceContextProvider>
         </HttpServiceContextProvider>
       </AuthDataServiceContextProvider>
     </BrowserRouter>
