+++ frontend/src/services/http-service.tsx
@@ -1,47 +1,15 @@
-import { AuthApi, Configuration } from 'msadoc-client';
+import { Configuration } from 'msadoc-client';
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import { firstValueFrom } from 'rxjs';
 import { AjaxConfig } from 'rxjs/ajax';
 
 import { ENVIRONMENT } from '../env';
-import { LoginHttpResponse, UnknownHttpError } from '../models/api';
-import { APP_ROUTES } from '../routes';
-
-import {
-  AccessAndRefreshToken,
-  useAuthDataServiceContext,
-} from './auth-data-service';
 
 interface HttpService {
-  /**
-   * Login using the given username and password.
-   * If the login succeeds, the returned auth/refresh tokens will automatically be stored.
-   */
-  performLogin: (
-    username: string,
-    password: string,
-  ) => Promise<LoginHttpResponse | UnknownHttpError>;
-
-  refreshAuthToken(): Promise<AccessAndRefreshToken | undefined>;
-
   createConfiguration(accessToken?: string): Configuration;
   getErrorStatus(error: unknown): number | undefined;
 }
 
-interface State {
-  isSignedIn: boolean;
-}
-
 function useHttpService(): HttpService {
-  const navigate = useNavigate();
-  const authDataService = useAuthDataServiceContext();
-
-  const [state, setState] = React.useState<State>({
-    isSignedIn:
-      authDataService.state.accessAndRefreshToken?.refreshToken !== undefined,
-  });
-
   function createConfiguration(accessToken?: string): Configuration {
     if (accessToken === undefined) {
       return new Configuration({
@@ -74,80 +42,6 @@ function useHttpService(): HttpService {
     });
   }
 
-  async function performLogin(
-    username: string,
-    password: string,
-  ): Promise<LoginHttpResponse | UnknownHttpError> {
-    try {
-      const response = await firstValueFrom(
-        new AuthApi(createConfiguration()).authControllerLogin({
-          loginRequestDto: {
-            username: username,
-            password: password,
-          },
-        }),
-      );
-
-      authDataService.setAccessAndRefreshToken({
-        accessToken: response.access_token,
-        refreshToken: response.refresh_token,
-      });
-      return {
-        status: 200,
-        data: response,
-      };
-    } catch (error) {
-      const errorStatus = getErrorStatus(error);
-      return {
-        status: errorStatus === 401 ? 401 : 0,
-        data: undefined,
-      };
-    }
-  }
-
-  /**
-   * Use the Refresh Token to generate a new Auth Token.
-   */
-  async function refreshAuthToken(): Promise<
-    AccessAndRefreshToken | undefined
-  > {
-    if (
-      authDataService.state.accessAndRefreshToken?.refreshToken === undefined
-    ) {
-      navigate(APP_ROUTES.login);
-      return undefined;
-    }
-
-    try {
-      const response = await firstValueFrom(
-        new AuthApi(
-          createConfiguration(
-            authDataService.state.accessAndRefreshToken.refreshToken,
-          ),
-        ).authControllerRefreshToken({
-          refreshTokenRequestDto: {
-            refresh_token:
-              authDataService.state.accessAndRefreshToken.refreshToken,
-          },
-        }),
-      );
-
-      authDataService.setAccessAndRefreshToken({
-        accessToken: response.access_token,
-        refreshToken: response.refresh_token,
-      });
-      return {
-        accessToken: response.access_token,
-        refreshToken: response.refresh_token,
-      };
-    } catch (error) {
-      // In the future, we might want to distinguish cases like "the client currently has no internet connection" and "the token is invalid". However, the following should be fine for now.
-      authDataService.deleteAccessAndRefreshToken();
-      navigate(APP_ROUTES.login);
-      return undefined;
-    }
-  }
-
   /**
    * Get the "status" field of an error object returned when a http request fails.
    */
@@ -165,30 +59,7 @@ function useHttpService(): HttpService {
     return status;
   }
 
-  // Whenever the auth token are deleted from local storage, update the state accordingly.
-  React.useEffect(() => {
-    setState({
-      isSignedIn:
-        authDataService.state.accessAndRefreshToken?.refreshToken === undefined,
-    });
-  }, [authDataService.state.accessAndRefreshToken]);
-
-  const TOKEN_REFRESH_INTERVAL_MS = 60000;
-  // Whenever signed in, refresh the auth tokens regularly.
-  React.useEffect(() => {
-    const interval = setInterval(() => {
-      if (!state.isSignedIn) {
-        return;
-      }
-      void refreshAuthToken();
-    }, TOKEN_REFRESH_INTERVAL_MS);
-
-    return () => clearInterval(interval);
-  });
-
   return {
-    performLogin: performLogin,
-    refreshAuthToken: refreshAuthToken,
     createConfiguration: createConfiguration,
     getErrorStatus: getErrorStatus,
   };
